# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TicketAPI",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Registra um novo usuário",
                "operationId": "a3f252b10cbf8e270739c3ada00c6042",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "username",
                                    "email",
                                    "birth_date",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "joaosilva"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "birth_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senhaSegura123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senhaSegura123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário registrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos na requisição"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Autentica um usuário",
                "operationId": "eaf28e54bd4a4126b9668fe26e6d1411",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "senhaSegura123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abc123..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/v1/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoga o token atual do usuário autenticado",
                "operationId": "4f953f79b68161bb5120e6f95126c000",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Logout realizado com sucesso."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token inválido ou ausente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Lista tickets filtrando por status (abertos ou fechados)",
                "operationId": "5b6909c0a1de09b0287a11d6aec0dd92",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtra os tickets por status: 'abertos' ou 'fechados'",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "abertos",
                                "fechados"
                            ],
                            "example": "abertos"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de tickets retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Ticket"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parâmetro de status inválido"
                    },
                    "403": {
                        "description": "Acesso negado. Role necessária: admin|atendente"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Cria um novo ticket",
                "operationId": "d9cabc2f38b61406ec12f4d930b66cd7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Erro ao salvar formulário"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Nada acontece ao clicar em salvar"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "ti"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Usuário não possui permissão para abrir tickets"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}/assign": {
            "patch": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Atribui um ticket a um usuário",
                "operationId": "a5dabfb5f2a16371d27ce52528d8d957",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "assigned_user_id"
                                ],
                                "properties": {
                                    "assigned_user_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atribuído com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Responsável atribuído com sucesso."
                                        },
                                        "ticket": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ticket não encontrado"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tickets/{id}/close": {
            "patch": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Fecha um ticket",
                "operationId": "2433a0d141986623768b7edc129b1dc7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do ticket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket fechado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ticket fechado com sucesso."
                                        },
                                        "ticket": {
                                            "$ref": "#/components/schemas/Ticket"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ticket já está fechado"
                    },
                    "404": {
                        "description": "Ticket não encontrado"
                    },
                    "403": {
                        "description": "Acesso negado. Role necessária: admin|atendente"
                    },
                    "401": {
                        "description": "Usuário não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Ticket": {
                "title": "Ticket",
                "required": [
                    "id",
                    "title",
                    "description",
                    "type",
                    "created_by"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Erro no sistema"
                    },
                    "description": {
                        "type": "string",
                        "example": "O sistema está travando quando clico em salvar."
                    },
                    "type": {
                        "type": "string",
                        "example": "ti"
                    },
                    "opened_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-28T12:34:56"
                    },
                    "closed_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    },
                    "assigned_user_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "created_by": {
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "Usuário",
                "required": [
                    "id",
                    "full_name",
                    "username",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "full_name": {
                        "type": "string",
                        "example": "João da Silva"
                    },
                    "username": {
                        "type": "string",
                        "example": "joaosilva"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "joao@example.com"
                    },
                    "birth_date": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-01T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-05-20T14:45:00Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "PersonalAccessToken",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Ticket",
            "description": "Ticket"
        }
    ]
}